requirements.txt
fastapi
uvicorn[standard]
python-dateutil
sqlalchemy
